fix: countBy 반환 타입 Int로 수정

+++ src/main/kotlin/com/weeth/domain/board/domain/repository/BoardRepository.kt
@@ -36,7 +36,7 @@ interface BoardRepository :
     fun countByClubIdAndTypeNotAndIsDeletedFalse(
         clubId: Long,
         type: BoardType,
-    ): Long
+    ): Int
 
     fun existsByClubIdAndNameAndIsDeletedFalse(
         clubId: Long,
